EditorInvisible берём из core
