+++ package.json
@@ -24,7 +24,7 @@
     "metadata-postgres": "2.0.23-beta.2",
     "metadata-pouchdb": "2.0.23-beta.2",
     "paper-jsdom-canvas": "0.12.8",
-    "windowbuilder": "github:oknosoft/windowbuilder-core#9ab6200be901a502c7c3337b9582d52a0d555950"
+    "windowbuilder": "github:oknosoft/windowbuilder-core#7b23455b94550d5248d16d8466380da41855f6a9"
   },
   "devDependencies": {
     "md5-file": "^5.0",
